Rename "Expected points" to "Target points"
This clearly led to some confusion when I shared it; we don't "expect" a
team to score this many points, it's the number of points that a typical
championship team would score from this match.

Renaming this to "target points" hopefully clarifies that (which solves
the problem where the 19th place team playing the 20th place team, both
teams have 3 "expected" points, which is just bizarre. Now they'll both
have 3 "target" points, which is more reasonable)